+++ packages/content/types/vuex.d.ts
@@ -1,7 +1,7 @@
-import { NuxtContentInstance } from '.'
+import { QueryBuilder } from "./query-builder";
 
-declare module 'vuex/types/index' {
+declare module "vuex/types/index" {
   interface Store<S> {
-    $content: NuxtContentInstance
+    $content: QueryBuilder;
   }
 }
